requirements.txt
numpy==1.23.4
Pillow==9.3.0
pycocotools==2.0.6
torch==1.13.0
torchvision==0.14.0
